move even size column to raw components
